projects/cybertools/requirements.txt
openai==1.3.6
python-dotenv==1.0.0
httpx==0.25.0